fix: added support for strings containing `\0`
Previously, `Character.Any.Char` was assigned to a default value of `\0`.
This meant that it cannot be differentiated from `new Character('\0')`.

My suggested change is to make `Character.Char` nullable, and base pattern-matching decisions on the nullability instead.

I have added a test to showcase the lack of support for '\0' before the change.